# break reference cycle in inspect_span_from_stack