@dneuhaeuser-zalando 2020-10-29
FYI We have patched opentracing_utils in one of our own applications and this is the impact:

<img width="1879" alt="Screenshot 2020-10-29 at 11 01 11" src="https://user-images.githubusercontent.com/37899626/97553794-2dfa3480-19d6-11eb-8f29-4238c62c2051.png">