@dneuhaeuser-zalando 2021-03-01
@vetinari Could you take a look at this? We currently run a patched version and it would be nice to get rid of the patch.