@pitr 2021-03-03
v0.20.1 has been released - https://pypi.org/project/opentracing-utils/

Thanks you for following up on this.